v3.1.0(书写即将完成，近日发布)

+++ SguanFOC库v3.1.0(无感foc，浮点...writing)/Sguan_Hall.c
@@ -69,7 +69,7 @@ void Hall_Loop(HALL_STRUCT *hall){
 
     // 4.霍尔数据处理并输出角度值
     Sector = (Signal_a << 2) | (Signal_b << 1) | (Signal_c);
-    hall->go.Output_Rad = (v[Sector - 1]*Value_2PI)/6.0f;
+    hall->go.Output_Rad = (v[Sector - 1]*(float)Value_2PI)/6.0f;
 }
